requirements.txt
# ASR Model: Whisper
git+https://github.com/openai/whisper.git

# Torch backend (required by Whisper)
torch

# Audio preprocessing
ffmpeg-python
soundfile
numpy

# Evaluation metrics
jiwer

# FastAPI Web Backend
fastapi
uvicorn
python-multipart

# Trigger Word Detection with Vosk
vosk
sounddevice

# CORS support
aiofiles

pyttsx3
pyaudio

streamlit 
requests

websocket-client

webrtcvad

